Moved sink outside internal to make it public
