+++ cmd/humanlog/main.go
@@ -15,8 +15,8 @@ import (
 	types "github.com/humanlogio/api/go/types/v1"
 	"github.com/humanlogio/humanlog"
 	"github.com/humanlogio/humanlog/internal/pkg/config"
-	"github.com/humanlogio/humanlog/internal/pkg/sink/stdiosink"
 	"github.com/humanlogio/humanlog/internal/pkg/state"
+	"github.com/humanlogio/humanlog/sink/stdiosink"
 	"github.com/mattn/go-colorable"
 	"github.com/urfave/cli"
 )
