+++ sink/sink.go
