+++ sink/stdiosink/colors.go
